created controller and exception handler

+++ SpacehubRepository/src/main/java/com/project/spacehub/repository/BookingRepository.java
@@ -12,6 +12,6 @@ import com.project.spacehub.models.Booking;
  *
  */
 
-public interface BookingRepository extends JpaRepository<Booking, Long> {
+public interface BookingRepository extends JpaRepository<Booking, Integer> {
 
 }
