+++ SpacehubRepository/src/main/java/com/project/spacehub/repository/ProductRepository.java
@@ -8,11 +8,14 @@ import org.springframework.stereotype.Repository;
 
 import com.project.spacehub.models.Product;
 
+import java.util.Optional;
+
 /**
  * @author gbemisola
  *
  */
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long>{
+public interface ProductRepository extends JpaRepository<Product, Integer>{
+
 
 }
